permissions-policy/historical.html

<!DOCTYPE html>
<title>Historical Feature Policy features should not be present</title>
<link rel="help" href="https://www.w3.org/TR/permissions-policy/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
// Test 1: document.featurePolicy API should not exist (replaced by document.permissionsPolicy)
test(function() {
    assert_false('featurePolicy' in document,
                 'document.featurePolicy should not exist');
}, 'document.featurePolicy should not be present (use document.permissionsPolicy instead)');

// Test 2: HTMLIFrameElement.featurePolicy API should not exist (replaced by permissionsPolicy)
test(function() {
    assert_false('featurePolicy' in HTMLIFrameElement.prototype,
                 'HTMLIFrameElement.prototype.featurePolicy should not exist');
}, 'HTMLIFrameElement.prototype.featurePolicy should not be present');

// Test 3: Feature-Policy header should be ignored when Permissions-Policy is present
// The header file sets: Feature-Policy: sync-xhr 'none'
// And: Permissions-Policy: sync-xhr=*
// If Feature-Policy is correctly ignored, sync XHR should work
test(function() {
    var xhr = new XMLHttpRequest();
    xhr.open("GET", document.location.href, false);
    xhr.send();
    assert_equals(xhr.status, 200, 'Sync XHR should succeed because Permissions-Policy allows it');
}, 'Feature-Policy header should be ignored when Permissions-Policy is present');

// Test 4: iframe allow attribute should work with Permissions Policy
async_test(function(t) {
    var iframe = document.createElement('iframe');
    iframe.src = 'resources/historical-iframe.html';
    iframe.allow = "sync-xhr 'none'";

    window.addEventListener('message', t.step_func_done(function(e) {
        if (e.source !== iframe.contentWindow) return;
        assert_equals(e.data.result, 'blocked',
                      'sync-xhr should be blocked in iframe when allow="sync-xhr \'none\'"');
    }));

    document.body.appendChild(iframe);
}, 'iframe allow attribute blocks features using Permissions Policy syntax');
</script>
</body>
